aiModel/requirements.txt
Package versions:-

torch: 2.0.1+cpu
torchvision: 0.15.2+cpu
ultralytics: 8.3.225
Pillow: 12.0.0
numpy: 1.26.4